requirements.txt
# Streamlit Engineering Drawing Extractor Dependencies

# Core application framework
streamlit>=1.28.0

# AWS SDK and services
boto3>=1.34.0
botocore>=1.34.0

# PDF processing
PyPDF2>=3.0.1
PyMuPDF>=1.23.0  # fitz for advanced PDF handling
Pillow>=10.0.0   # PIL for image processing

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Testing framework
pytest>=7.4.0
pytest-mock>=3.11.0

# Type hints and validation
typing-extensions>=4.7.0

# Configuration and environment management
python-dotenv>=1.0.0

# Logging and utilities
python-dateutil>=2.8.2